j4a - removed HW read iomask
It was just adding a needless extra stage of delay on reads.
Phrases like: ` $80 im! 4 io@ ` can be replaced with ` 4 io@ $80 and `
which is more standard for forth anyway.
(Where `: im! $8000 io! ;` is the io mask preset)
Still uses the highest bit, which is slow (takes longer to load as
an immediate, since any immediate with that bit cleared is one
instruction ), and still clears with any subsequent io op *from the same
core*.
And is still very useful for writing only particular bits in port/dir or
leds without clobbering the bits you don't want to change.
Read-modify-write isn't safe, because of the fine-threaded hardware
multitasking implicit in the j4a.