increased il to 21.4 bar
Due to the 'low' level being higher than usual. The il limit is supposed
to cause hesitation in case retraction isn't finished, so it should be
high enough to do so. If a shot is fired before retraction, the inlet
valve can be broken. Ideally this should detect the rapid rise at the
moment of retraction completion, rather than just a tuned threshold.